+++ core/src/avm2/globals/flash/events/mouse_event.rs
@@ -40,8 +40,8 @@ pub fn update_after_event<'gc>(
 pub(super) fn local_to_stage_x<'gc>(
     activation: &mut Activation<'_, 'gc>,
     this: Object<'gc>,
-    slot_x: u32,
-    slot_y: u32,
+    slot_x: usize,
+    slot_y: usize,
 ) -> Result<Value<'gc>, Error<'gc>> {
     if let Some(evt) = this.as_event() {
         let local_x = this.get_slot(slot_x).coerce_to_number(activation)?;
@@ -66,8 +66,8 @@ pub(super) fn local_to_stage_x<'gc>(
 pub(super) fn local_to_stage_y<'gc>(
     activation: &mut Activation<'_, 'gc>,
     this: Object<'gc>,
-    slot_x: u32,
-    slot_y: u32,
+    slot_x: usize,
+    slot_y: usize,
 ) -> Result<Value<'gc>, Error<'gc>> {
     if let Some(evt) = this.as_event() {
         let local_x = this.get_slot(slot_x).coerce_to_number(activation)?;
